feat: Add a `codelists check` command
This checks the state of the codelist on disk against what's specified
in codelists.txt and what's recorded as being downloaded in
codelists.json.

There's some temporary code which spots that it's being run in a Github
Action and falls back to using the old verification method (which checks
against OpenCodelists) so that existing repos don't fail immediately
under the new method.

Closes #12