feat: Add upgrade command and update available message
This is just a helpful wrapper around running pip upgrade step, but its
more user friendly

Additionally, we nag user to run the new upgrade command if we detect
a more recent of opensafely is available. Not quite auto-update, but
might get us quite far as is.

Note: we cache the latest version for 2 hours, to avoid spamming pypi on
every command.