chore: refactor argparse code to allow use of parse_known_args()
The new jupyter command wants to accept any args, to add to the jupyter
invocation. However, that cannot be specified when using
`parser_args()`, as all `--option` args must be understood by argparse,
only positional onces can be var args.

Calling parse_known_args() allows the unknown ones to be collected
separatey, but is uneeded for most commands, and the looser parsing is
less user friendly.

So, allow each subcommand to opt in to using the looser parsing.