fix: jupyter another attempt at dealing with tty in git-bash
The goal is to automatically wrap our call to "docker run -it ..." in
winpty if we detect we are running in git-bash's broken default
terminal.

Both cmd.exe and Windows Terminal provide a proper tty so do not need
this.

The problem is that passing the `-v=...` paths as arguments to winpty
forces them through git-bash path translation, which we had previously
been able to avoid as we were going via subprocess.run(), and the path
translation mucked up the -v mappings.

So, this change does two thinks:

1) try to detect when were are in the severely limit git-bash/mingw
   environement, and wrap the docker command in winpty only then.

2) Escape the docker side of the mount from git-bash with
   `://workspace`, and revert to using standard backslash paths on the
   host side.

This has been manually tested in git-bash and in cmd.exe and worked on
a least one windows machine.