chore: refactor vendoring tooling
The current vendoring tooling is broken:

 - it allows independant updates of job-runner versus job-runner
 dependencies, so things can get out of sync

 - it does not use the same version pins as jobrunner for job-runners
   dependencies, which is bad, but also stuff breaks on more recent
   requests, so it blocks updates

We fix this by simplifying the tooling that updates both in lockstep,
using the versions pinned in the appropriate version of job-runner.

In addition to the, upcoming new dependencies require inspecting
installed packages via `pkg_resources`. By default, the `vendoring` tool
deletes all the package metadata required for this to work.

To fix this, we:

a) have forked vendoring and added support to not delete the packaging
metadata. This works and will be submitted upstream as a PR.

b) have changed the packaging to include these as package data

c) manually add the `opensafely/_vendor` directory to the places
pkg_resources will look for package metadata on start up.

As a result of this, we've moved to using a more recent `vendoring`
version. We were on 0.3.3, but vendoring is on 1.2, 6 releases on.  So
this has brought us up to and added some improvements - it shouts a bit
more about some some things, and looks prettier.