feat: update jobrunner to v2.59.11
Using the new tooling, this also updates some existing deps that had
fallen behind job-runner (requests and pytdantic and their deps) and
also adds the new opentelemetry deps.

Notice that it includes the dist-info dirs as per previous commit

The reliance on protobufs is unfortunately. OTel standard supports json
serializaion, but the python sdk does not yet. When it does, we can
hopefully drop the protobuf dependencies.