feat: Add `opensafely exec` command
This is for running ad-hoc OpenSAFELY actions. As an illustration,
running this command:

    opensafely exec foo:v1 do_something --a=1

Should do roughly the same thing as this action in `project.yaml`:

    my_action:
      run: foo:v1 do_something --a=1

The key difference is that rather than having `needs` and `outputs`
control which files are available, we mount the entire workspace in so
the command can read and write any files it likes.

This is motivated particularly by a desire to remove friction from the
Data Builder tutorial. Being able to run commands, see log and error
output immediately, and read and create arbitrary files provides a much
smoother on-ramp than forcing everything through `project.yaml`.

It also means we can support commands like Data Builder's
`create-dummy-tables` which are development tools that don't fit
naturally into `project.yaml` at all.

Note that this command is only usable with images that have been rebuilt
using the latest base-docker image, following this change:
https://github.com/opensafely-core/base-docker/pull/17

There are various areas for improvement and expansion here which I
haven't attempted to tackle at this stage:

 * Adding an `--entrypoint` argument to allow customising the image
   entrypoint for debugging purposes.
 * Passing the appropriate product key for the `stata-mp` image.
 * Setting the `OPENSAFELY_BACKEND` environment variable to
   `expectaions` at the local job-runner does.