fix: support console interactivity in exec
This extracts and enhances the ensure_tty function used in the jupyter
command to be used with `opensafely exec`.

In doing so, we make some changers:

a) move from `platform.system()` to `sys.platform`, as the former calls
subprocess.run, which messes with our run fixture expectations, and we
do not need anything beyond what sys.platform provides.

b) expand RunFixture to explicitly handle run calls that we expect to be
wrapped in winpty. This assumes the Windows tests run with git-bash
installed, which is True in Github actions at least.