fix: Add new utils package with run_docker function
The function handles.

 - running an opensafely image
 - with the right args
 - with a directory mounted at /workspace
 - will run as the invoking user on linux by default
 - can handle interactivity and git-bash/winpty

We then use this new function in juypter command. This means it now runs
as the user rather than root, which was an issue when running in linux.

Also, we move from `platform.system()` to `sys.platform`, as the former
calls subprocess.run, which messes with our run fixture expectations,
and we do not need anything beyond what sys.platform provides.