Add a rudimentary dataset definition check
This checks tables, based on the presence of accessing table attributes.

This is rudimentary, but parallels what is present for study
definitions.

This code could be refactored and improved more. But I started doing
that and ended up down a very big rabbit hole where the code was not
necessarily much clearer, particularly when reviewing the changes. As
this code should be considered temporary, and the changes here are
relatively small compared with the original code, it is perhaps better
to leave it as is for now.

The differences in formatting between cohort-extractor and ehrQL also
make this a little bit gnarly.

That said, the existing repeated opening of the same files might warrant
later refactoring, although it's unlikely that this code is particularly
performance critical.