Fix version used in packaging tests
The default contents of the `opensafely/VERSION` file used to be a
string which was (as well as being an invalid version specifier)
consiered lower than all our published versions. There's a comment to
this effect in one of the tests, which expects that the `upgrade`
command will therefore have something to do.

However, in commit 83dda1359c5c this changed and the default contents
became `99.99.99` i.e. _higher_ than every published version. This meant
that the test no longer tested the upgrade process. That's fortunate
because had it worked the Docker test that comes after it would have
been testing the latest published version, not the version under test.

We now do some horrible monkey business to temporarily change the
version file during the packaging tests. I did try just setting the
default to a low version, but that triggers a "there's a new version"
prompt which breaks a bunch of the tests.