Explicitly mock open_browser call in test_rstudio.py
The call to `webbrowser.open` calls a subprocess, which will fail as
we've not called run.expect for it. However, this error was hidden as
test exits quickly enough that it doesn't appear, unless its slow for
some reason.

This change mocks the call to open_browser, and checks that it is
actually called.