Refactor SubprocessRunFixture to handle concurrent declarations
The original implementation requires that commands were executed in the
order they were declared, which has worked fine so far. However, when
testing some of our threaded behaviour, this was not a valid assumption.

This refactors the fixture to add support for setting `run.concurrent
= True`, which will match *any* specified command, rather than the next
one. The default behaviour is still to require them in order.

As an implementation detail, it meant switching away from deque to list
made sense, as we want to now be able to remove from anywhere in the
list.

Note that tests using this new functionality will be in next commit.