feat: refactor rstudio and jupyter into launch command
They are very similar anyway, and share a lot of code already, and both
start a server and open a browser.

However but the main motivation was to be able to use the same pattern
as `opensfely exec image`. This is cleaner, easier to learn and allows
for more control over version in the future.