feat: Add flag to launch to run in background
This is useful for running in codespaces.

Because functional testing of forground process is complex, previous
tests for launch have used the `conftests.run` fixture to stub out
expected calls to docker.

However, the addition of the --background flag allows for
proper functional tests of the feature, as they are more tractable.

So, this change copies over the docker cleanup functions from
job-runner, and adds functional tests using the background. This tests
only run on linux, as we don't have docker in CI on macos/linux. But in
addition to testing the new --background flag, they provide better
assurance that opensafely launch is working.

I have left the version testing as stubbed tests for now, as that is a)
mainly about argument parsing, and much faster than the functional
tests.