Improve idempotency for re-running launch, and add --force
If a tool already running, launch the browers pointing at that. The
default naming means you can run one tool per directory.

Adds the --force option to require a full restart. This is useful if you
want to ensure the latest version of an image.

The motivation for this is better UX in codespaces, where we will
reconnect to a running instance by default.

The implementation for --force is a little messy do to docker semantics,
so we test it functionally, even though its slower.