fix: add --list-project-images flag to info command.
This is information is useful in codespaces to know what runtime version
to enable support for for the current project. The `opensafely pull`
command already has this code, but exposing it there didn't make much
sense. So I added to the info command.

In the process, I added tests for the base info command, as there were
none. This flag is hidden as it is intended for use by developers, not
end users.