Try using default ports before choosing random ports
In many cases, the default port for jupyter (8888) and rstudio (8787)
will be free, and using them makes the ports consistant and match user's
expectations.

Additionally, in codespaces, when manually running rstudio:v2, choosing
a random port every time fills up the PORTS tab over time with old
defunct ports. So defaulting to the same port each time is better there.

Because codespaces is currently runs its own instance of RStudio (v1) on
8787, we default to 8788 if we detect we are in a codespace. This will
also allow us to label that port as "RStudio (v2)", which will help
differentiate it from the v1 version.

As a fallback, if that port is busy, we still use a random port. The
launch command is already idempotent, so it will re-use an already
running instance anyway. We would only fall back to a random port if the
thing using that port was *not* run by the opensafely launch command. In
practice, I think this means we will almost never use a random port.

Also, realised we could use context managers with socket.socket(), so I
did.