Handle draft codelists
Previously it was nearly impossible to add a draft codelist to
codelists.txt as going to the URL in a browser would redirect to that
codelist's builder URL, which would not pass the URL pattern validation.

Copying the draft codelist URL from a user's codelists page and then
adding this to codelist.txt manually or via `codelists add` bypasses
this block.

When a csv download of a draft codelist is requested from opencodelists
it returns the HTML page for that codelist along with text stating its
draft status, which the CLI would willingly put into a csv file in the
codelists directory.

We don't want this behaviour and we don't want users to add draft
codelists to projects as they are still mutable at this point.

This change checks for draft codelists and prevents their download.