Fix tracing test fixture issue
So, conftest.py is special. It is loaded but *not* imported in to
sys.modules.

When we imported get_trace from it, it was re-imported.  This reset the
test trace set up, and caused some issues.

This change moves the long-lived gloabl test exporter to live in
tests/jobrunner/fixtures.py, which is imported normally, i.e. once.

This is a latent bug that pre-existed in the jobrunner code base, but
for some reason, we never hit. We hit it here I think because it is not
in the root tests/conftest.py file, like in jobrunner proper, but in
tests/jobrunner/conftests.py

Have created issue to fix in job-runner https://github.com/opensafely-core/job-runner/issues/1181