Skip docker tests on windows in CI
Also, just_executable() on Windows returns a window path. However, we
are still using `sh` in our justfiles, it is installed on Github
windows images.  So it cannot execute the result of that function.

Instead, we simply rely on just being on the path, and use a direct
`just` rather than `{{ just_executable() }}`.

In theory, this could break if just is not on path, but its hard to
think of a when that would happen.