Formally remove the otel deps from prod dependencies
All except opentelemetry-api, which is still vendored.

The actual code has already been deleted - this commit has the changes
that effected that removal, for easier understanding.

The setuptools package cannot be vendored, which is why it is our only
formal dependency. We needed to tweak the new requirements-prod just
method to not try to add it, by passing `--no-allow-unsafe`. We cannot
get rid of a runtime setup tools dependency just yet, but we maybe can
in future by upgrading otel libraries.

This meant we can also remove the pkg_resources hack we needed for otel
deps to work when vendored.