Fix jobrunner and tests to work with the new dependency reality
This tries to import the optional dependencies, and ensures the code
handles when they are not available. It adds an explict
TRACING_AVAILABLE flag, and if it's False, does no-ops in our manual
tracing functions.

It turns out we cannot mix and match otel libraries imported from our
vendored lib and globally in pip. This is an unfortunante implementation
detail of the otel implementation, the module path affects some internal
identity checks on objects.

So, we have our tracing module provide an abstraction. It either imports
the whole otel stack from the global namespace, if found (like in dev
and when installed with opensafely[tracing], but falls back to importing
the core `trace` module from `opensafely._vendor.opentelemetry` if they
are not found.

The rest of our code imports the `trace` module from jobrunner.tracing,
so it is always consistent in which version it uses. This actually
resulted in a much clearer api usage.