Add otel dependencies as a package extra and test
This modifies the package metadata with an extra called "tracing", which
installs the right packages.

It adds a functional test that installs into a virtualenv using this
extra, and that if you run a job with tracing configured, tracing
does work.

Also, I expanded the existing installation test, that now effectively
installs without the dependencies, with a similar functional test to
ensure that jobs still run in a normal installation. This caught a bug,
which is fixed by adding another no-op when TRACING_AVAILABLE is False.