Ensure dev deps defer the production versions
We actually had different versions of some shared deps in tests.

The fact that we're vendoring makes this awkward, but less brittle that
otherwise, given it means we can have 2 versions, and you explicitly
import them on different module paths.

But it's more consistant to ensure everything's on the same version.