Upgrade the opentelemetry dependencies.
We can only go up to 1.33.1 (1.15 for exporters), as 1.34 dropped python
3.8 support.

However, it does mean we can drop setuptools as formal dependency. This
is great, as we are back to zero dependencies, and it was displaying
a pkg_resources deprecation warning for many users.

The updated version uses importlib-metadata to inspect what other
opentelemetry packages installed. Unfortunately, there is a bug in
1.33.1, it doesn't correctly handle the case when there are none.

It uses entrypoints, that opentelemetry-api has an entrypoint, so this
is normally not a problem, as there will always be 1. However, because
we are vendoring it, this entrypoint never gets installed/registered.
So, we had to manually add the entrypoint in our package metadata,
pointing to the vendored package's entrypoint.

It's worth noting that testing these upgrades is made much easier by the
tests that install the package in a clean venv, with and without the
otel deps, and run and actual job.