feat: Reimplement how we upgrade opensafely on windows.
Changes in the way more recent setuptools build the windows entrypoints
means the way we previously did this no longer worked.

The new method creates a background subprocess to do the upgrade, then
exit, so that pip is free to overrite the opensafely.exe entrypoint.

As this is critical functionality, have enhnaced the tests a bit
explicitly check the version after upgrading. It's already got multiple
functional tests on all supported platforms.