Handle publish during upgrade functional tests.
Post merge, we run the tests and publish CI actions in parallel

This means that the current version can be changed in the middle of the
functional upgrade tests.

So change the point when we check the expected version in the test to
reduce the race window, and attempt to cope with the possiblity of
publish mid-test.

Ideally, we'd make depend on tests, which serializes them, and is
argubly the correct thing to do anyway. But that requires re-working the
CI quite a bit, so just doing this fix for now.