Remove unused permissions file
This code which used to request this file was removed on 2025-10-03:
 * https://github.com/opensafely-core/opensafely-cli/pull/364

However, older versions of the `opensafely` tool would continue to
attempt to fetch it so we left it in place for a while to avoid
triggering confusing errors and support requests.

By this point, all users have had ample time to upgrade (which they will
have been prompted to do by the tool itself) and so if this _does_ cause
issues it's evidence of users ignoring upgrade prompts which is itself
useful to know about.

(Note, there's no need to trigger a new release here: the client fetched
the file directly from `main` on Github.)